crash-tests: prevent the process from getting stuck when there's not enough entropy
The update of the `rand` crate changed the entropy requirements for the
test binary used by the crash-tests, and now the binary gets stuck
indefinetely waiting for entropy to generate the random file name. Fixed
by simply making `/dev/random` a symlink to `/dev/urandom` in the test
virtual machine.